--- teams/TailwindTraders-M365/service/.env
@@ -1,2 +1,0 @@
-BotId=4916b306-b8ca-4c37-89d0-5e5757097c6c
-BotPassword=q.-NdhKZMs-6R98B1GA6tx~gPpllV4w-2i
