+++ teams/TailwindTraders-M365/service/Learn Together Bot/.env
@@ -1,0 +1,2 @@
+BotId=e9286a48-0b71-4d85-ad2b-32a2614781b6
+BotPassword=P86~ugsDlZuMRe1mp~-4_3i4YllF-s_ge~
